requirements.txt
# Core Framework
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.8
langchain-core>=0.1.0

# AI/ML Libraries
sentence-transformers>=2.6.1
torch>=2.2.0
tensorflow>=2.16.1
huggingface-hub>=0.20.0

# Vector Database
chromadb>=0.4.0

# Database Connectivity
pymysql>=1.0.2
mysql-connector-python>=8.0.33
sqlalchemy>=2.0.0

# Environment Management
python-dotenv>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
regex>=2023.0.0
